+++ src/Resources/app/administration/src/component/nosto-job-status-badge/component.js
@@ -4,6 +4,7 @@
 
 import template from './nosto-job-status-badge.html.twig';
 
+/** @private */
 export default {
     template,
